refactor: Moved dataclasses to their own module
